refactor: move components into own directory
