+++ src/components/SocialNetworks/SocialNetworks.tsx
@@ -1,0 +1,30 @@
+import React, { FunctionComponent } from 'react';
+import { useTranslation } from 'react-i18next';
+
+import './SocialNetworks.scss';
+import { socialNetworks } from './list';
+
+export interface SocialNetworksProps {}
+
+const SocialNetworks: FunctionComponent<SocialNetworksProps> = () => {
+  const { t } = useTranslation();
+
+  return (
+    <section className="social-networks">
+      <h2 className="social-networks__heading">
+        {t('social_networks_heading')}
+      </h2>
+      <ul className="social-networks__list">
+        {socialNetworks.map(({ href, name }) => (
+          <li key={name} className="social-networks__list-item">
+            <a href={href} target="_blank" rel="noopener noreferrer">
+              {name}
+            </a>
+          </li>
+        ))}
+      </ul>
+    </section>
+  );
+};
+
+export default SocialNetworks;
